move rlvr to utils file
