+++ compose_rl/data/__init__.py
@@ -11,12 +11,18 @@ from compose_rl.data.preference_data import (
     pairwise_preference_dataset_collate_fn,
 )
 from compose_rl.data.prompt_data import prompt_dataset_collate_fn
+from compose_rl.data.rlvr_utils import (
+    extract_gsm8k_answer,
+    prepare_gsm8k_prompt,
+)
 
 __all__ = [
-    'finegrained_preference_dataset_collate_fn',
-    'pairwise_preference_dataset_collate_fn',
-    'prompt_dataset_collate_fn',
     'build_pairwise_preference_dataloader',
     'build_finegrained_preference_dataloader',
     'build_prompt_dataloader',
+    'extract_gsm8k_answer',
+    'finegrained_preference_dataset_collate_fn',
+    'pairwise_preference_dataset_collate_fn',
+    'prepare_gsm8k_prompt',
+    'prompt_dataset_collate_fn',
 ]
